capsules: move screen virtualizer to its own module
