+++ capsules/extra/src/lib.rs
@@ -116,4 +116,4 @@ pub mod tsl2561;
 pub mod usb;
 pub mod usb_hid_driver;
 pub mod virtual_kv;
-pub mod virtual_screen_split;
+pub mod virtualizers;
