+++ capsules/extra/src/virtualizers/mod.rs
@@ -1,0 +1,5 @@
+// Licensed under the Apache License, Version 2.0 or the MIT License.
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+// Copyright Tock Contributors 2025.
+
+pub mod screen;
